requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-cov==7.0.0
httpx==0.25.1
python-multipart==0.0.6
